refactoring / bug fix

+++ main.c
@@ -9,19 +9,18 @@ int main() {
     size_t n;
     scanf("%zd", &n);
 
-    vectorVoid v = createVectorV(0, sizeof(int));
+    vectorVoid v = createVectorV(0, sizeof(float));
     for (int i = 0; i < n; i++) {
-        int x;
-        scanf("%d", &x);
+        float x;
+        scanf("%f", &x);
 
         pushBackV(&v, &x);
     }
-
     for (int i = 0; i < n; i++) {
-        int x;
+        float x;
         getVectorValueV(&v, i, &x);
 
-        printf("%d ", x);
+        printf("%f ", x);
     }
 
     return 0;
